requirements.txt
pygame==2.5.2
supabase==2.3.0
bcrypt==4.0.1
requests==2.31.0
GitPython==3.1.40